requirements.txt
aiofiles
aiohttp
beautifulsoup4
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
httpx
motor
numpy
opencv-python
pillow
psutil
py-tgcalls==2.2.8
pytgcrypto
kurigram
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
unidecode
uvloop==0.21.0; sys_platform != "win32"
yt-dlp
youtube-search
youtube-search-python
py-yt-search

